requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.1
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytesseract  # You also need Tesseract-OCR installed on system
six==1.17.0
tabula-py==2.10.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
